Redirecting to Success or Error

+++ Pages/Index.cshtml.cs
@@ -12,11 +12,13 @@ namespace UploadFilesToBB05022023.Pages
 {
     public class IndexModel : PageModel
     {
+        //temporary token, needs to be changed for every session. Will have to look into OAuth 2
         static string token =
-           "sl.BYWuXnavEGaRs5TVnYL0Qheg5vA41LpxdfkpkHOQO4OrMivSUktotecVsti7VoEUUZVPBOWSg_eOFAislAy0JU9Hx4kW2iIkNHr-mAR7HbmdwjdL_LWydFg8DJFILIrckVvBOcE";
+           "sl.BYXKJ6J_drfFeqRb3wewomKkTCwcBtFwFp7ZspDK4HgM2ZW2NXcD5ycaWqVf9b19o528VmO-YHzK7m-ayexl0MAva2nZw9J3W9tSvIVqDkefvUWyvjzNmet-YjmoxAfhhZug61A";
 
-            Dropbox.Api.DropboxClient dbx;
-        public bool Connected { get; private set; }
+        Dropbox.Api.DropboxClient dbx;
+
+        private bool Connected = false;
         private bool isConnecting = false;
         private bool errorOccured = false;
         string errorMessage;
@@ -24,65 +26,78 @@ namespace UploadFilesToBB05022023.Pages
         public void OnGet()
         {
         }
-        public void OnPost(IFormFile file)
-        {
-            if (errorOccured||isConnecting) 
-                return;
 
+        public IActionResult OnPost(IFormFile file)
+        {           
             if (!Connected)
             {
                 ConnectToDB();
             }
-            UploadFile(file);           
+            
+            return !errorOccured && UploadFile(file).Result
+             ? RedirectToPage("Success")
+             : RedirectToPage("Error"); //I'd like to send errorMessage as well, but that'll have to wait.
         }
 
         async void ConnectToDB()
         {
             if (!Connected && !isConnecting)
             {
+                isConnecting = true;
                 try
                 {
-                    isConnecting = true;
                     dbx = new DropboxClient(token);
                     var id = await dbx.Users.GetCurrentAccountAsync();
                 }
-                catch(Exception ex)
+                catch (Exception ex)
                 {
                     errorOccured = true;
-                    errorMessage = ex.Message;                                     
+                    errorMessage = ex.Message;
                 }
-                    isConnecting = false;
+
+                isConnecting = false;
             }
         }
 
-        async void UploadFile(IFormFile file)
+        async Task<bool> UploadFile(IFormFile file)
         {
-            try
+            //file can't be directly loaded to DB bc they use different types of file
+            bool success = false;
+
+            if (file is null)
             {
-                //incoming file cannot be directly uploaded to Dropbox
-                //because they use different File classes...
-
-                var dbPath = "/UploadedFiles/" + file.FileName;
-                var localPath = "UploadedFiles\\" + file.FileName;
-
-                await using var temoraryFile = new FileStream(localPath, FileMode.Create);
-                await file.CopyToAsync(temoraryFile);
-                temoraryFile.Close();
-
-                var mem = new MemoryStream(System.IO.File.ReadAllBytes(localPath));
-                await dbx.Files.UploadAsync(dbPath, body: mem);
-
-                if (System.IO.File.Exists(localPath))
+                errorMessage = "no file chosen";
+            }
+            else
+            {
+                try
                 {
-                    System.IO.File.Delete(localPath);
+                    var dbPath = "/UploadedFiles/" + file.FileName;
+                    var localPath = "UploadedFiles\\" + file.FileName;
+
+                    //copy file to folder
+                    await using var temoraryFile = new FileStream(localPath, FileMode.Create);
+                    await file.CopyToAsync(temoraryFile);
+                    temoraryFile.Close();
+
+                    //upload file to Dropbox
+                    var mem = new MemoryStream(System.IO.File.ReadAllBytes(localPath));
+                    await dbx.Files.UploadAsync(dbPath, body: mem);
+
+                    //delete file from folder
+                    if (System.IO.File.Exists(localPath))
+                    {
+                        System.IO.File.Delete(localPath);
+                    }
+                    success = true;
+                }
+                catch (Exception ex)
+                {
+                    errorMessage = ex.Message;
                 }
             }
-            catch(Exception ex)
-            {
-                errorOccured = true;
-                errorMessage = ex.Message;
-            }
-        }
 
+            return success;
+        }       
     }
 }
